SCB-321 Optimize zipkin plugin.
